+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.1.1] - 2019-10-01
 ### Fixes
 - Added error reporting for when the user calls a command they don't have permissions for.
 - Fixed a bug which caused the reminder date to be formatted incorrectly when confirming reminders.
 
 ### Database
-- Refactored the database connection so that it pings to stay alive. 
+- Refactored the database connection so that it keeps itself alive. 
 
 ## [4.1.0] - 2019-09-23
 ### Role Flairing